Update #10
Made a few adjustments to the config folder but I am still having trouble with the null movement for some reason every time I start up TF2 and join a server it just doesn't register the null movement script. Also testing if the "hold f to inspect weapon" is disabled.